Add statement. New tests

+++ src/Esketit.API/ResponseModels/AccountSummaryResponse.cs
@@ -21,21 +21,20 @@
 		public object[] pendingWithdrawals { get; set; }
 		public object[] unconfirmedWithdrawals { get; set; }
 		public decimal interestPaid { get; set; }
-		public int interestBonusPaid { get; set; }
+		public decimal interestBonusPaid { get; set; }
 		public decimal bonusPaid { get; set; }
 		public decimal referralBonusPaid { get; set; }
 		public decimal secondaryMarketIncome { get; set; }
 		public decimal secondaryMarketExpense { get; set; }
 		public decimal secondaryMarketTotal { get; set; }
 		public decimal totalIncome { get; set; }
-		public int internalRateOfReturnPercent { get; set; }
+		public decimal internalRateOfReturnPercent { get; set; }
 		public bool internalRateOfReturnPercentCalculated { get; set; }
 		public decimal principalCurrent { get; set; }
 		public decimal principalLate15 { get; set; }
 		public decimal principalLate30 { get; set; }
 		public decimal principalLate60 { get; set; }
-		public int principalLate61plus { get; set; }
+		public decimal principalLate61plus { get; set; }
 		public int totalInvestments { get; set; }
 	}
-
 }
